refactor: move registry to factory file and add initializer
