+++ ext/resourcemgr/factory_test.go
