+++ go.mod
@@ -67,6 +67,8 @@ require (
 
 require go.uber.org/multierr v1.7.0
 
+require github.com/mitchellh/mapstructure v1.4.1
+
 require (
 	cloud.google.com/go v0.94.0 // indirect
 	cloud.google.com/go/storage v1.16.1 // indirect
@@ -131,7 +133,6 @@ require (
 	github.com/mitchellh/colorstring v0.0.0-20190213212951-d06e56a500db // indirect
 	github.com/mitchellh/copystructure v1.2.0 // indirect
 	github.com/mitchellh/go-testing-interface v1.0.0 // indirect
-	github.com/mitchellh/mapstructure v1.4.1 // indirect
 	github.com/mitchellh/reflectwalk v1.0.2 // indirect
 	github.com/muesli/termenv v0.9.0 // indirect
 	github.com/oklog/run v1.0.0 // indirect
